fix(parser): widen emphasis run combine before u8 overflow
CM rule 9/10 in the emphasis resolver combines the open run and close
run lengths to check the multiple-of-3 carve-out. Both sides are u8;
on a 200-asterisk attack input the intermediate addition panics in
debug builds with "attempt to add with overflow".

Cast each u8 to usize before the add so the combined length is
computed in a 64-bit register. No behavioral change on inputs whose
total fits in u8.

Surfaced by a 200-asterisk deep-nesting probe.

+++ dmc-parser/src/inline.rs
@@ -33,7 +33,9 @@ pub(crate) fn resolve_emphasis_delims(out: &mut Vec<Node>, delims: &mut [DelimRe
       }
       // CM rule 9/10: skip when both delim sides are open+close and the
       // combined length is a multiple of 3 (unless this side is itself).
-      let combined = (d.run + delims[i].run) as usize;
+      // Cast each `u8` to `usize` before adding so a 200-`*` adversarial
+      // run does not panic on the intermediate `u8 + u8` overflow.
+      let combined = (d.run as usize) + (delims[i].run as usize);
       let both_open_close = (d.can_open && d.can_close) || (delims[i].can_open && delims[i].can_close);
       if both_open_close && combined.is_multiple_of(3) && !(d.run as usize).is_multiple_of(3) {
         continue;
